Adds moveable gzofstreams/gzifstreams.
Currently gzifstreams and gzofstreams are not moveable,
as is the default behavior of istreams and ostreams.
This behavior only exists because the basic_ios class is
not moveable for reasons covered here:

https://stackoverflow.com/a/20774666/1371191

Because all gzifstreams and gzofstreams use the same
underlying buffer, there is nothing preventing them from
being moveable.This enables clearer ownership semantics
for these streams and direct use in standard library containers.
To achieve this currently, access is required through raw/smart
pointers.

The one caveat is that the move constructors and assignment
operators are declared noexcept for containers requiring
nothrow move constructible types. Some functions called by
the default constructor (which is called in the move constructor)
are not marked noexcept. Afaict, there isn't a situation where
a throw would happen during move construction unless something
wonky happened in setting locales. This exception would be
expected to be thrown during default initialization of a
gzifstream/gzofstream anyway.